Bug fixes in Test Event Utility

+++ datamodels/altimeter.go
@@ -2,9 +2,9 @@ package datamodels
 
 // This struct represents a single altimeter sensor reading
 type AltimeterReading struct {
-	EventState     int `json:"EventState"`
-	AltimeterState int `json:"AltimeterState"`
-	Altitude       int `json:"Altitude"`
+	EventState     int     `json:"EventState"`
+	AltimeterState int     `json:"AltimeterState"`
+	Altitude       float64 `json:"Altitude"`
 }
 
 // This struct represents a single altimeter sensor event
